[core] split headless_view.cpp into implementation-specific files
also adds a EAGL implementation for iOS
